Exception Handler 통합 및 정적 팩토리 메서드 네이밍 변경_#54
