+++ src/main/java/io/oopy/coding/common/config/security/SecurityConfig.java
@@ -43,7 +43,7 @@ public class SecurityConfig {
             "/api/v1/contents/get", "/api/v1/comments/get",
             "/api/v1/category/get", "/api/v1/mark/get",
             "/api/v1/profile/**",
-            "/login/oauth2/**", "/api/v1/feed/title", "/api/v1/feed/body"
+            "/login/oauth2/**", "/api/v1/feed/title", "/api/v1/feed/body", "/api/v1/feed/nickname"
     };
 
     @Bean
