+++ src/main/java/io/oopy/coding/common/security/filter/JwtAuthenticationFilter.java
@@ -56,7 +56,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
             "/login/oauth2/**",
             "/api/v1/contents/get", "/api/v1/comments/get",
             "/favicon.ico",
-            "/api/v1/feed/title", "/api/v1/feed/body",
+            "/api/v1/feed/title", "/api/v1/feed/body", "/api/v1/feed/nickname",
             "/api/v1/category/get", "/api/v1/mark/get"
     );
