+++ src/main/java/io/oopy/coding/domain/content/dto/ContentDTO.java
@@ -31,7 +31,7 @@ public class ContentDTO {
                 .views(content.getViews())
                 .repoName(content.getRepoName())
                 .repoOwner(content.getRepoOwner())
-                .publish(content.getComplete())
+                .publish(content.getPublish())
                 .contentImageUrl(content.getContentImageUrl())
                 .createdAt(content.getCreatedAt())
                 .updatedAt(content.getUpdatedAt())
