+++ src/main/java/io/oopy/coding/domain/content/entity/ContentCategory.java
@@ -28,7 +28,7 @@ public class ContentCategory extends Auditable {
     @Column(name = "category_type", nullable = false)
     private String type;
 
-    public static ContentCategory fromContentAndCategory(Content content, Category category) {
+    public static ContentCategory newInstance(Content content, Category category) {
         return ContentCategory.builder()
                 .content(content)
                 .category(category)
